style: split classes out into dedicated files
per PR request
