+++ Corgibytes.Freshli.Cli/Functionality/Cache.cs
@@ -2,37 +2,9 @@ using System;
 using System.IO;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
-using ServiceStack;
 
 namespace Corgibytes.Freshli.Cli.Functionality
 {
-    [Index(nameof(Key), IsUnique = true)]
-    public class CachedProperty
-    {
-        public int Id { get; set; }
-        public string Key { get; set; }
-        public string Value { get; set; }
-    }
-
-    public class CacheContext : DbContext
-    {
-        public static readonly DirectoryInfo DefaultCacheDir = new DirectoryInfo(Environment.GetEnvironmentVariable("HOME") + "/.freshli");
-        public static DirectoryInfo CacheDir = DefaultCacheDir;
-
-        private static readonly string cacheDbName = "freshli.db";
-        public string DbPath { get; }
-
-        public DbSet<CachedProperty> CachedProperties { get; set; }
-
-        public CacheContext()
-        {
-            DbPath = Path.Join(CacheDir.ToString(), cacheDbName);
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder options)
-            => options.UseSqlite($"Data Source={DbPath}");
-    }
-
     public static class Cache
     {
         private static void MigrateIfPending(CacheContext context)
